requirements.txt
streamlit==1.10.0
pandas==1.4.2
numpy==1.21.5
scikit-learn==1.0.2
joblib==1.1.0
tensorflow-cpu==2.9.1


